fix one more case

+++ src/rl_str_proc.cc
@@ -67,7 +67,7 @@ StrProc::get(int start, int end)
 
   string tmp;
   for (size_t i = (size_t) start;
-       (i < end) && (i < _str_coll.size()); ++i) {
+       (i < (size_t) end) && (i < _str_coll.size()); ++i) {
     tmp += _str_coll[i] + " ";
   }
   return tmp.substr(0,tmp.length()-1);
